+++ src/modules/board/board.controller.ts
@@ -23,8 +23,8 @@ import { FileInterceptor } from '@nestjs/platform-express';
 import { UserJwtAuthGuard } from '../auth/guards/user-jwt.guard';
 import { UserId } from '../../decorators/user-id.decorator';
 import { UserCreateResultInterface } from '../../interfaces/user-create-result.interface';
-import { PageOptionsDto } from '../../global/common/dto/page-options.dto';
-import { PaginationResponseDto } from '../../global/common/dto/pagination-response.dto';
+import { PageOptionsDto } from '../../global/common/paginate/dto/offset-paginate/page-options.dto';
+import { PaginationResponseDto } from '../../global/common/paginate/dto/pagination-response.dto';
 import { Board } from './entity/board.entity';
 import { GetBoardDto } from './dto/get-board.dto';
 import { BoardStatus } from '../../types/enums/boardStatus.enum';
