+++ src/modules/board/board.service.ts
@@ -8,9 +8,9 @@ import { S3Service } from '../../config/s3/s3.service';
 import { RedisService } from '../../config/redis/redis.service';
 import { BoardMapper } from './mapper/board.mapper';
 import { UserCreateResultInterface } from '../../interfaces/user-create-result.interface';
-import { PageOptionsDto } from '../../global/common/dto/page-options.dto';
-import { PaginationResponseDto } from '../../global/common/dto/pagination-response.dto';
-import { PageMetaDto } from '../../global/common/dto/page-meta.dto';
+import { PageOptionsDto } from '../../global/common/paginate/dto/offset-paginate/page-options.dto';
+import { PaginationResponseDto } from '../../global/common/paginate/dto/pagination-response.dto';
+import { PageMetaDto } from '../../global/common/paginate/dto/offset-paginate/page-meta.dto';
 import { PageNotExists } from '../../global/exception/pageException/Page-Not-Exists-Exception';
 import { GetBoardDto } from './dto/get-board.dto';
 import { BoardStatus } from '../../types/enums/boardStatus.enum';
